requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis


# Developer Installed Packages #
# ------------------------------------------------------------------------------
beautifulsoup4==4.9.3
cssutils==2.3.0 

# SEARCH INDEX IMPORTS & PLUGINS
# ------------------------------------------------------------------------------
django-haystack==3.0
whoosh==2.7.4


# Admin
# ------------------------------------------------------------------------------
django-jet2==1.0.12
FeedParser==5.2.1
python-magic==0.4.15
filemagic==1.6
libmagic==1.0
pyxlsb==1.0.3
django-ipware==3.0.1
django-autocomplete-light==3.8.2
django-tinymce==3.3.0
django-filebrowser-no-grappelli==3.8.0
google-api-python-client==1.4.1
django-admin-honeypot==1.1.0

# Auto shell import in django
# ------------------------------------------------------------------------------
django-shell-plus==1.1.7

# Minify HTML in django
# ------------------------------------------------------------------------------
# Todo: find the uses of these to minify wesbites and others to improve security
django-htmlmin==0.11.0
python-memcached==1.59

# List of countries
# ------------------------------------------------------------------------------
django-countries-plus==1.3.1


# Add image resize for models
# ------------------------------------------------------------------------------
django-resized==0.3.11


# Add geolocation to django app
# ------------------------------------------------------------------------------
geoip2==4.1.0
maxminddb==2.0.3

django-pipeline==2.0.6

# Blog post readtime
# ------------------------------------------------------------------------------
readtime==1.1.1


# Dynamic Scraper for django
# ------------------------------------------------------------------------------
# Django Dynamic Scraper (DDS) is an app for Django which builds on 
# top of the scraping framework Scrapy and lets you create and manage 
# Scrapy spiders via the Django admin interface. It was originally 
# developed for german webtv program site 
# ------------------------------------------------------------------------------
# django-dynamic-scraper==0.13.3 


# Dynamic Scraper for django
# ------------------------------------------------------------------------------
# Scrapy is a fast high-level web crawling and web scraping framework, 
# used to crawl websites and extract structured data from their pages. 
# It can be used for a wide range of purposes, from data mining to 
# monitoring and automated testing.
# ------------------------------------------------------------------------------
Scrapy==2.5.0 
lxml==4.6.3 
scrapy-djangoitem==1.1.1
word2word==1.0.0 
requests-html==0.10.0
django-filter==2.4.0 

#Todo: Django Filter Search, Ajax Commenting, Language Integration, Cookie Consent
# --extra-index-url https://dl.fontawesome.com/TOKEN/fontawesome-pro/python/simple/
fontawesome-free==5.15
django-tables2==2.4.0


# Blog requirements
django-category==2.0.1 #
django-comments-dab==2.6.0
django-next-prev==1.1.0


# White noise for static files
whitenoise==5.3.0